+++ packages/frontend/src/pages/Dashboard.tsx
@@ -18,7 +18,6 @@ import MantineIcon from '../components/common/MantineIcon';
 import DashboardDeleteModal from '../components/common/modal/DashboardDeleteModal';
 import DashboardDuplicateModal from '../components/common/modal/DashboardDuplicateModal';
 import { DashboardExportModal } from '../components/common/modal/DashboardExportModal';
-import DashboardFiltersWarningModal from '../components/common/modal/DashboardFiltersWarningModal';
 import Page from '../components/common/Page/Page';
 import SuboptimalState from '../components/common/SuboptimalState/SuboptimalState';
 import DashboardFilter from '../components/DashboardFilter';
@@ -179,9 +178,6 @@ const Dashboard: FC = () => {
             (tile) => tile.type === DashboardTileTypes.SEMANTIC_VIEWER_CHART,
         );
     }, [dashboardTiles]);
-
-    const [isFilterWarningModalOpen, setIsFilterWarningModalOpen] =
-        useState(false);
 
     // tabs state
     const [activeTab, setActiveTab] = useState<DashboardTab | undefined>();
@@ -560,20 +556,14 @@ const Dashboard: FC = () => {
 
     const handleEnterEditMode = useCallback(() => {
         resetDashboardFilters();
-        setIsFilterWarningModalOpen(false);
-        history.replace({
-            pathname: `/projects/${projectUuid}/dashboards/${dashboardUuid}/edit`,
-            search: '',
+        // Defer the redirect
+        void Promise.resolve().then(() => {
+            history.replace({
+                pathname: `/projects/${projectUuid}/dashboards/${dashboardUuid}/edit`,
+                search: '',
+            });
         });
     }, [history, projectUuid, dashboardUuid, resetDashboardFilters]);
-
-    const handleEditModeClicked = useCallback(() => {
-        if (haveFiltersChanged) {
-            setIsFilterWarningModalOpen(true);
-        } else {
-            handleEnterEditMode();
-        }
-    }, [haveFiltersChanged, handleEnterEditMode]);
 
     if (dashboardError) {
         return <ErrorState error={dashboardError.error} />;
@@ -694,20 +684,34 @@ const Dashboard: FC = () => {
                         onExport={exportDashboardModalHandlers.open}
                         setAddingTab={setAddingTab}
                         onTogglePin={handleDashboardPinning}
-                        onEditClicked={handleEditModeClicked}
+                        onEditClicked={handleEnterEditMode}
                     />
                 }
                 withFullHeight={true}
             >
                 <Group position="apart" align="flex-start" noWrap px={'lg'}>
-                    {dashboardChartTiles && dashboardChartTiles.length > 0 && (
-                        <DashboardFilter
-                            isEditMode={isEditMode}
-                            activeTabUuid={activeTab?.uuid}
-                        />
-                    )}
+                    <Box style={{ flexGrow: 1, overflow: 'auto' }}>
+                        {/* This Group will take up remaining space (and not push DateZoom) */}
+                        <Group
+                            position="apart"
+                            align="flex-start"
+                            noWrap
+                            px={'lg'}
+                        >
+                            {dashboardChartTiles &&
+                                dashboardChartTiles.length > 0 && (
+                                    <DashboardFilter
+                                        isEditMode={isEditMode}
+                                        activeTabUuid={activeTab?.uuid}
+                                    />
+                                )}
+                        </Group>
+                    </Box>
+                    {/* DateZoom section will adjust width dynamically */}
                     {hasDashboardTiles && !hasNewSemanticLayerChart && (
-                        <DateZoom isEditMode={isEditMode} />
+                        <Box style={{ marginLeft: 'auto' }}>
+                            <DateZoom isEditMode={isEditMode} />
+                        </Box>
                     )}
                 </Group>
                 <Flex style={{ flexGrow: 1, flexDirection: 'column' }}>
@@ -755,13 +759,6 @@ const Dashboard: FC = () => {
                         uuid={dashboard.uuid}
                         onClose={duplicateModalHandlers.close}
                         onConfirm={duplicateModalHandlers.close}
-                    />
-                )}
-                {isFilterWarningModalOpen && (
-                    <DashboardFiltersWarningModal
-                        onConfirm={handleEnterEditMode}
-                        onClose={() => setIsFilterWarningModalOpen(false)}
-                        opened={isFilterWarningModalOpen}
                     />
                 )}
             </Page>